# server/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "GBCC Example Swagger autogen",
    "description": "Auto generated documentation for GBCC App",
    "version": "1.0.0"
  },
  "host": "gbcc.dix.si:9080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/games/": {
      "get": {
        "description": "List all games.",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/gamespaginated/{page}/{pagesize}": {
      "get": {
        "description": "List all games. Ordered by pages of size pagesize",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagesize",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/game/{gamename}": {
      "get": {
        "description": "Search a game by name. The name can be a pattern",
        "parameters": [
          {
            "name": "gamename",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/games": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/games/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/players/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/playerspaginated/{page}/{pagesize}": {
      "get": {
        "description": "List all players. Ordered by pages of size pagesize",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagesize",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/player/{searchname}": {
      "get": {
        "description": "Search for a player the searchname will serach for pattern in First name and Last name",
        "parameters": [
          {
            "name": "searchname",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/playedgamesaginated/{playerid}/{page}/{pagesize}": {
      "get": {
        "description": "List all games played by player. (defined by ID) Ordered by pages of size pagesize",
        "parameters": [
          {
            "name": "playerid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagesize",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/player": {
      "post": {
        "description": "Insert a new player",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "borndate": {
                  "example": "any"
                },
                "documentnr": {
                  "example": "any"
                },
                "gdpraccept": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/player/{id}": {
      "put": {
        "description": "Update a players name or last name",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "delete a player",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/playedgames/": {
      "get": {
        "description": "List all games played",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/playagame": {
      "post": {
        "description": "Record that a player had played a game",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "playerid": {
                  "example": "any"
                },
                "gameid": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}